Redo program structure
