+++ benchmark_observer_llm/benchmark/__init__.py
@@ -1,0 +1,7 @@
+import logging
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(levelname)s %(asctime)s %(message)s",
+    datefmt="%Y-%m-%dT%H:%M:%S",
+)
